+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `hidden` property to the facets query.
+
+### Changed
+- `productClusterIds` don't show up in breadcrumb anymore.
+
 ## [1.2.0] - 2020-05-29
 
 ### Changed